Transaction view : now displaying fractions correctly, added SENDER field. Config page : fixed inability to change textbox values (there was a handler firing spuriously and updating them and overiding user input) Log page : corrected auto-scroll to move to latest log entry in view.

+++ ExscudoTestnetGUI/commitClass.cs
@@ -15,7 +15,7 @@ namespace ExscudoTestnetGUI
         {
 
             [JsonProperty("amount")]
-            public ulong Amount { get; set; }
+            public decimal Amount { get; set; }
 
             [JsonProperty("recipient")]
             public string Recipient { get; set; }
